ls: revert previous patch, use getdents().
We can use sbrk() if we don't use opendir.  So use getdents, which
isn't quite equivalent to the previous buffered read of 16 bytes at a
time, but we use the same buffer, so it's as close as possible.

Before:
UNIX ls:
2.39 2.39 2.39
UNIX ls -l:
4.24 4.24 4.24

After:
UNIX ls:
2.33 2.36 2.34
UNIX ls -l:
4.18 4.18 4.17